+++ sdk/textanalytics/azure-ai-textanalytics/tests/test_repr.py
@@ -26,6 +26,20 @@ def text_document_statistics():
     return model, model_repr
 
 @pytest.fixture
+def request_statistics():
+    model = _models.RequestStatistics(
+        documents_count=1,
+        valid_documents_count=1,
+        erroneous_documents_count=0,
+        transactions_count=1
+    )
+
+    model_repr = "RequestStatistics(documents_count=1, valid_documents_count=1, erroneous_documents_count=0, transactions_count=1)"
+
+    assert repr(model) == model_repr
+    return model, model_repr
+
+@pytest.fixture
 def text_analytics_warning():
     model = _models.TextAnalyticsWarning(
         code="LongWordsInDocument",
@@ -441,49 +455,52 @@ class TestRepr():
         assert error.code == "UnsupportedLanguageCode"
         assert error.message == "Supplied language not supported. Pass in one of: de,en,es,fr,it,ja,ko,nl,pt-PT,zh-Hans,zh-Hant"
 
-    def test_analyze_batch_actions_result_recognize_entities(self, recognize_entities_result):
+    def test_analyze_batch_actions_result_recognize_entities(self, recognize_entities_result, request_statistics):
         model = _models.AnalyzeBatchActionsResult(
             document_results=[recognize_entities_result[0]],
+            statistics=request_statistics[0],
             is_error=False,
             action_type=_models.AnalyzeBatchActionsType.RECOGNIZE_ENTITIES,
             completed_on=datetime.datetime(1, 1, 1)
         )
 
         model_repr = (
-            "AnalyzeBatchActionsResult(document_results=[{}], is_error={}, action_type={}, completed_on={})".format(
-                recognize_entities_result[1], False, "recognize_entities", datetime.datetime(1, 1, 1)
+            "AnalyzeBatchActionsResult(document_results=[{}], is_error={}, action_type={}, completed_on={}, statistics={})".format(
+                recognize_entities_result[1], False, "recognize_entities", datetime.datetime(1, 1, 1), request_statistics[1]
             )
         )
 
         assert repr(model) == model_repr
 
-    def test_analyze_batch_actions_result_recognize_pii_entities(self, recognize_pii_entities_result):
+    def test_analyze_batch_actions_result_recognize_pii_entities(self, recognize_pii_entities_result, request_statistics):
         model = _models.AnalyzeBatchActionsResult(
             document_results=[recognize_pii_entities_result[0]],
+            statistics=request_statistics[0],
             is_error=False,
             action_type=_models.AnalyzeBatchActionsType.RECOGNIZE_PII_ENTITIES,
             completed_on=datetime.datetime(1, 1, 1)
         )
 
         model_repr = (
-            "AnalyzeBatchActionsResult(document_results=[{}], is_error={}, action_type={}, completed_on={})".format(
-                recognize_pii_entities_result[1], False, "recognize_pii_entities", datetime.datetime(1, 1, 1)
+            "AnalyzeBatchActionsResult(document_results=[{}], is_error={}, action_type={}, completed_on={}, statistics={})".format(
+                recognize_pii_entities_result[1], False, "recognize_pii_entities", datetime.datetime(1, 1, 1), request_statistics[1]
             )
         )
 
         assert repr(model) == model_repr
 
-    def test_analyze_batch_actions_result_extract_key_phrases(self, extract_key_phrases_result):
+    def test_analyze_batch_actions_result_extract_key_phrases(self, extract_key_phrases_result, request_statistics):
         model = _models.AnalyzeBatchActionsResult(
             document_results=[extract_key_phrases_result[0]],
+            statistics=request_statistics[0],
             is_error=False,
             action_type=_models.AnalyzeBatchActionsType.EXTRACT_KEY_PHRASES,
             completed_on=datetime.datetime(1, 1, 1)
         )
 
         model_repr = (
-            "AnalyzeBatchActionsResult(document_results=[{}], is_error={}, action_type={}, completed_on={})".format(
-                extract_key_phrases_result[1], False, "extract_key_phrases", datetime.datetime(1, 1, 1)
+            "AnalyzeBatchActionsResult(document_results=[{}], is_error={}, action_type={}, completed_on={}, statistics={})".format(
+                extract_key_phrases_result[1], False, "extract_key_phrases", datetime.datetime(1, 1, 1), request_statistics[1]
             )
         )